Project/requirements.txt
asgiref==3.7.2
astroid==2.4.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2020.6.20
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.1
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
distlib==0.3.7
Django==4.2.6
djangorestframework==3.15.1
et-xmlfile==1.1.0
filelock==3.12.4
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.8
numpy==1.24.4
openpyxl==3.1.2
pbr==6.0.0
pillow==10.3.0
pipenv==2023.10.20
platformdirs==3.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pygame==2.1.2
Pygments==2.17.2
pylint==2.6.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyOpenSSL==24.1.0
PyPDF2==3.0.1
redis==5.0.4
requests==2.31.0
service-identity==24.1.0
six==1.15.0
soupsieve==2.5
sqlparse==0.4.4
stevedore==5.1.0
toml==0.10.2
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==2.0.7
virtualenv==20.24.6
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
wrapt==1.12.1
zope.interface==6.3
